+++ __test__/fixtures/type-alias-to-interface-heritage/old/index.ts
@@ -1,0 +1,5 @@
+export interface RequiredBase {
+  required: number;
+}
+
+export type Options = {};