+++ __test__/fixtures/type-alias-to-interface-incompatible/old/index.ts
@@ -1,0 +1,3 @@
+export type Cfg = {
+  a: number;
+};