+++ __test__/fixtures/type-alias-to-interface-index-sig-noop/old/index.ts
@@ -1,0 +1,3 @@
+export type Dict = {
+  [k: string]: number;
+};